Replace new file
