+++ test/numToKorean.test.ts
@@ -1,4 +1,4 @@
-import { numToKorean } from '../src/new-num-to-korean';
+import { numToKorean } from '../src/num-to-korean';
 
 describe('숫자 금액을 한글로 변환하기 테스트', () => {
   test('예외 값 처리', () => {
